+++ frontend/components/Sidebar.tsx
@@ -1,60 +1,15 @@
 "use client";
 import Image from "next/image";
 import { createContext, useState } from "react";
-import Cal from "../../frontend/app/assets/Calendar.png";
-import Chart from "../../frontend/app/assets/Chart.png";
-import Chart_fill from "../../frontend/app/assets/Chart_fill.png";
-import Chat from "../../frontend/app/assets/Chat.png";
 import arrow from "../../frontend/app/assets/control.png";
 import logo from "../../frontend/app/assets/logo.png";
-import Search from "../../frontend/app/assets/Search.png";
-import user from "../../frontend/app/assets/User.png";
 import SidebarItem from "./SidebarItem";
 
 export const SidebarContext = createContext<any>(false);
 
-const Sidebar = () => {
+const Sidebar = ({ MENUS }: any) => {
   const [open, setOpen] = useState(true);
-  const MENUS = [
-    { title: "Dashboard", src: Chart_fill },
-    {
-      title: "Inventory",
-      src: Chat,
-      submenu: [
-        { title: "View inventory", route: "/pages/hospital/Inventory" },
-        { title: "Inventory Statistics", route: "/pages/hospital/Inventory" },
-      ],
-    },
-    {
-      title: "Orders",
-      src: user,
-      submenu: [
-        { title: "View Orders", route: "/pages/hospital/Orders" },
-        { title: "Request New Order", route: "/pages/hospital/Inventory" },
-        { title: "Orders status", route: "/pages/hospital/Inventory" },
-        { title: "Track your order", route: "/pages/hospital/Inventory" },
-      ],
-    },
-    {
-      title: "Testings",
-      src: Cal,
-      submenu: [
-        { title: "View Testings", route: "" },
-        { title: "Request New Testing", route: "" },
-        { title: "Testing Status", route: "" },
-      ],
-    },
-    {
-      title: "Vendors",
-      src: Search,
-      submenu: [
-        { title: "View Vendors", route: "" },
-        { title: "Market Place", route: "" },
-        { title: "Testing Status", route: "" },
-      ],
-    },
-    { title: "Analytics", src: Chart, route: "/Analytics" },
-  ];
+
   return (
     <div
       className={` ${
@@ -83,7 +38,7 @@ const Sidebar = () => {
         </h1>
       </div>
       <ul className="pt-6">
-        {MENUS.map((Menu, index) => (
+        {MENUS.map((Menu: any, index: any) => (
           <>
             <SidebarContext.Provider value={[open, setOpen]}>
               <SidebarItem Menu={Menu} index={index} />
